Release version 0.4.5 with libcurl 8.15.0.

+++ Sources/download-curl.sh
@@ -1,4 +1,18 @@
-curl --location --remote-name --progress-bar https://github.com/greatfire/curl-apple/releases/download/8.14.1/curl.xcframework.zip
+#!/usr/bin/env sh
+
+version=$1
+exptected_checksum=$2
+
+curl --location --remote-name --progress-bar "https://github.com/greatfire/curl-apple/releases/download/$version/curl.xcframework.zip"
+
+actual=$(shasum -a 256 curl.xcframework.zip | awk '{print $1}')
+
+if [ $actual != $exptected_checksum ]; then
+    echo "ERROR: Checksum verification failed."
+
+    exit 1
+fi
+
 rm -rf curl.xcframework
 unzip curl.xcframework.zip
 rm curl.xcframework.zip
